Chmnt prio

+++ serveur/Ot_Sims_Givebox/Ot_Sims_Givebox/Models/Prio.cs
@@ -9,7 +9,7 @@ namespace Ot_Sims_Givebox.Models
     public partial class Offre
     {
 
-        public int prio { get; set; }
+        public double prio { get; set; }
 
     }
